Moved some files from auth services to common services.
