+++ api/internal/common/id.go
@@ -1,7 +1,6 @@
-package firestore_services
+package common
 
 import (
-	cmn "Codex-Backend/api/internal/common"
 	"errors"
 	"math/rand"
 	"time"
@@ -16,7 +15,7 @@ func GenerateID(prefix string) (string, error) {
 	// Generate a ULID using the current timestamp and the entropy source
 	id, err := ulid.New(ulid.Timestamp(time.Now()), entropy)
 	if err != nil {
-		return "", &cmn.Error{Err: errors.New("ID Service Error - Generate ULID: " + err.Error())}
+		return "", &Error{Err: errors.New("Common Services Error - Generate ULID: " + err.Error())}
 	}
 
 	// Add a prefix to the generated ID
